add show, update, detail product

+++ server/controllers/categoryCtrl.ts
@@ -2,6 +2,7 @@ import { Request, Response } from 'express'
 import Category from '../models/Category'
 import { checkArrayEquality, getUniqueValues } from '../utils/helper'
 import { pagination } from '../utils/pagination'
+import Product from '../models/Product'
 
 const categoryCtrl = {
   create: async(req: Request, res: Response) => {
@@ -100,6 +101,14 @@ const categoryCtrl = {
         totalData: categoryCount,
         totalPage
       })
+    } catch (err: any) {
+      return res.status(500).json({ msg: err.message })
+    }
+  },
+  readAll: async(req: Request, res: Response) => {
+    try {
+      const category = await Category.find()
+      return res.status(200).json({ category })
     } catch (err: any) {
       return res.status(500).json({ msg: err.message })
     }
@@ -247,12 +256,13 @@ const categoryCtrl = {
     try {
       const { id } = req.params
       const category = await Category.findById(id)
-      // query product with currrent category
+      const product = await Product.find({ category: id })
       
       if (!category)
         return res.status(404).json({ msg: `Category with ID ${id} not found.` })
 
-      // if length of product is greater than 0, then delete operation can't be done, otherwise delete category
+      if (product.length > 0)
+        return res.status(400).json({ msg: `Several products still have ${category.name} category.` })
 
       await Category.findByIdAndDelete(id)
       return res.status(200).json({ msg: `Category with ID ${id} has been deleted successfully.` })
